Initial upload
In mostly functional state, but may be rough around the edges. New tableaus are brute-force computed with matrix inverses for now instead of pivoting. The hope was to permute the columns so that the basic and nonbasic variables are untangled, and to sort within each set in increasing order for readability and as an invariant to maintain. Unfortunately, the process of permuting the columns was less straightforward than I'd anticipated.